+++ game-frontend/src/app/deployed-units-big/deployed-units-big.component.ts
@@ -1,8 +1,8 @@
+import { Component, OnInit } from '@angular/core';
+
 import { ObtainedUnit } from '../shared-pojo/obtained-unit.pojo';
 import { BaseComponent } from '../base/base.component';
-import { UnitPojo } from '../shared-pojo/unit.pojo';
-import { Component, OnInit } from '@angular/core';
-import { UnitService } from 'app/service/unit.service';
+import { UnitService } from '../service/unit.service';
 
 @Component({
   selector: 'app-deployed-units-big',
@@ -18,6 +18,15 @@ export class DeployedUnitsBigComponent extends BaseComponent implements OnInit {
   }
 
   public ngOnInit() {
+    this._findInMyPlanet();
+  }
+
+  public onDeletion(): void {
+    delete this.obtainedUnits;
+    this._findInMyPlanet();
+  }
+
+  private _findInMyPlanet(): void {
     this.loginSessionService.findSelectedPlanet.filter(planet => !!planet).subscribe(planet => {
       this._unitService.findInMyPlanet(planet.id).subscribe(units => this.obtainedUnits = units);
     });